requirements.txt
beautifulsoup4
matplotlib
matplotlib-inline
numpy
openpyxl==3.0.10
pandas==1.5.2
pickleshare
Pillow
protobuf==3.20.3
python-dateutil
requests
scikit-learn==1.2.0
scipy==1.9.3
statsmodels==0.13.5
streamlit==1.17.0
streamlit-aggrid==0.3.3
streamlit-lottie==0.0.3
tinycss2
urllib3